+++ src/modules/Image/encodings/xcf.c
@@ -48,7 +48,7 @@ extern struct program *image_program;
 
 struct buffer
 {
-  unsigned int len;
+  size_t len;
   unsigned char *str;
 };
 
@@ -1162,7 +1162,7 @@ void image_xcf_f__decode_tiles( INT32 args )
       sp--;
     }
 
-    if( (unsigned)(tile->len) < (unsigned)(eheight * ewidth * bpp ))
+    if( (size_t)(tile->len) < (size_t)(eheight * ewidth * bpp ))
       error("Too small tile, was %d bytes, I really need %d\n",
             tile->len, eheight*ewidth * bpp);
